requirements.txt
openai>=0.27.0
requests
python-dotenv
